20085: Sharing dialog immediately saves changes
Added ALL_USERS as a sharing level distinct from PUBLIC.  Don't hide
the all users / public link in the permission list.  Switching between
PUBLIC/ALL_USERS/SHARED/PRIVATE have clearly distinct actions.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>

+++ src/store/sharing-dialog/sharing-dialog-actions.ts
@@ -31,12 +31,12 @@ import {
     ResourceObjectType
 } from "models/resource";
 import { resourcesActions } from "store/resources/resources-actions";
-import { getPublicGroupUuid } from "store/workflow-panel/workflow-panel-actions";
+import { getPublicGroupUuid, getAllUsersGroupUuid } from "store/workflow-panel/workflow-panel-actions";
 import { getSharingPublicAccessFormData } from './sharing-dialog-types';
 
 export const openSharingDialog = (resourceUuid: string, refresh?: () => void) =>
     (dispatch: Dispatch) => {
-        dispatch(dialogActions.OPEN_DIALOG({ id: SHARING_DIALOG_NAME, data: {resourceUuid, refresh} }));
+        dispatch(dialogActions.OPEN_DIALOG({ id: SHARING_DIALOG_NAME, data: { resourceUuid, refresh } }));
         dispatch<any>(loadSharingDialog);
     };
 
@@ -49,8 +49,8 @@ export const connectSharingDialogProgress = withProgress(SHARING_DIALOG_NAME);
 
 export const saveSharingDialogChanges = async (dispatch: Dispatch, getState: () => RootState) => {
     dispatch(progressIndicatorActions.START_WORKING(SHARING_DIALOG_NAME));
-    await dispatch<any>(savePublicPermissionChanges);
     await dispatch<any>(saveManagementChanges);
+    await dispatch<any>(savePublicPermissionChanges);
     await dispatch<any>(sendInvitations);
     dispatch(reset(SHARING_INVITATION_FORM_NAME));
     await dispatch<any>(loadSharingDialog);
@@ -133,7 +133,8 @@ const loadSharingDialog = async (dispatch: Dispatch, getState: () => RootState, 
             dispatch(snackbarActions.OPEN_SNACKBAR({
                 message: 'You do not have access to share this item',
                 hideDuration: 2000,
-                kind: SnackbarKind.ERROR }));
+                kind: SnackbarKind.ERROR
+            }));
             dispatch(dialogActions.CLOSE_DIALOG({ id: SHARING_DIALOG_NAME }));
         } finally {
             dispatch(progressIndicatorActions.STOP_WORKING(SHARING_DIALOG_NAME));
@@ -143,64 +144,86 @@ const loadSharingDialog = async (dispatch: Dispatch, getState: () => RootState, 
 
 export const initializeManagementForm = async (dispatch: Dispatch, getState: () => RootState, { userService, groupsService, permissionService }: ServiceRepository) => {
 
-        const dialog = getDialog<SharingDialogData>(getState().dialog, SHARING_DIALOG_NAME);
-        if (!dialog) {
-            return;
-        }
-        dispatch(progressIndicatorActions.START_WORKING(SHARING_DIALOG_NAME));
-        const resourceUuid = dialog?.data.resourceUuid;
-        const { items: permissionLinks } = await permissionService.listResourcePermissions(resourceUuid);
-        dispatch<any>(initializePublicAccessForm(permissionLinks));
-        const filters = new FilterBuilder()
-            .addIn('uuid', Array.from(new Set(permissionLinks.map(({ tailUuid }) => tailUuid))))
-            .getFilters();
+    const dialog = getDialog<SharingDialogData>(getState().dialog, SHARING_DIALOG_NAME);
+    if (!dialog) {
+        return;
+    }
+    dispatch(progressIndicatorActions.START_WORKING(SHARING_DIALOG_NAME));
+    const resourceUuid = dialog?.data.resourceUuid;
+    const { items: permissionLinks } = await permissionService.listResourcePermissions(resourceUuid);
+    dispatch<any>(initializePublicAccessForm(permissionLinks));
+    const filters = new FilterBuilder()
+        .addIn('uuid', Array.from(new Set(permissionLinks.map(({ tailUuid }) => tailUuid))))
+        .getFilters();
 
-        const { items: users } = await userService.list({ filters, count: "none", limit: 1000 });
-        const { items: groups } = await groupsService.list({ filters, count: "none", limit: 1000 });
+    const { items: users } = await userService.list({ filters, count: "none", limit: 1000 });
+    const { items: groups } = await groupsService.list({ filters, count: "none", limit: 1000 });
 
-        const getEmail = (tailUuid: string) => {
-            const user = users.find(({ uuid }) => uuid === tailUuid);
-            const group = groups.find(({ uuid }) => uuid === tailUuid);
-            return user
-                ? user.email
-                : group
-                    ? group.name
-                    : tailUuid;
-        };
-
-        const managementPermissions = permissionLinks
-            .filter(item =>
-                item.tailUuid !== getPublicGroupUuid(getState()))
-            .map(({ tailUuid, name, uuid }) => ({
-                email: getEmail(tailUuid),
-                permissions: name as PermissionLevel,
-                permissionUuid: uuid,
-            }));
-
-        const managementFormData: SharingManagementFormData = {
-            permissions: managementPermissions,
-            initialPermissions: managementPermissions,
-        };
-
-        dispatch(initialize(SHARING_MANAGEMENT_FORM_NAME, managementFormData));
-        dispatch(progressIndicatorActions.STOP_WORKING(SHARING_DIALOG_NAME));
+    const getEmail = (tailUuid: string) => {
+        const user = users.find(({ uuid }) => uuid === tailUuid);
+        const group = groups.find(({ uuid }) => uuid === tailUuid);
+        return user
+            ? user.email
+            : group
+                ? group.name
+                : tailUuid;
     };
 
+    const managementPermissions = permissionLinks
+        .map(({ tailUuid, name, uuid }) => ({
+            email: getEmail(tailUuid),
+            permissions: name as PermissionLevel,
+            permissionUuid: uuid,
+        }));
+
+    const managementFormData: SharingManagementFormData = {
+        permissions: managementPermissions,
+        initialPermissions: managementPermissions,
+    };
+
+    dispatch(initialize(SHARING_MANAGEMENT_FORM_NAME, managementFormData));
+    dispatch(progressIndicatorActions.STOP_WORKING(SHARING_DIALOG_NAME));
+};
+
 const initializePublicAccessForm = (permissionLinks: PermissionResource[]) =>
-    (dispatch: Dispatch, getState: () => RootState, ) => {
+    (dispatch: Dispatch, getState: () => RootState,) => {
+
+        const state = getState();
+
         const [publicPermission] = permissionLinks
-            .filter(item => item.tailUuid === getPublicGroupUuid(getState()));
-        const publicAccessFormData: SharingPublicAccessFormData = publicPermission
-            ? {
+            .filter(item => item.tailUuid === getPublicGroupUuid(state));
+
+        const [allUsersPermission] = permissionLinks
+            .filter(item => item.tailUuid === getAllUsersGroupUuid(state));
+
+        let publicAccessFormData: SharingPublicAccessFormData;
+
+        if (publicPermission) {
+            publicAccessFormData = {
                 visibility: VisibilityLevel.PUBLIC,
-                permissionUuid: publicPermission.uuid,
-            }
-            : {
-                visibility: permissionLinks.length > 0
-                    ? VisibilityLevel.SHARED
-                    : VisibilityLevel.PRIVATE,
-                permissionUuid: '',
+                initialVisibility: VisibilityLevel.PUBLIC,
+                permissionUuid: publicPermission.uuid
             };
+        } else if (allUsersPermission) {
+            publicAccessFormData = {
+                visibility: VisibilityLevel.ALL_USERS,
+                initialVisibility: VisibilityLevel.ALL_USERS,
+                permissionUuid: allUsersPermission.uuid
+            };
+        } else if (permissionLinks.length > 0) {
+            publicAccessFormData = {
+                visibility: VisibilityLevel.SHARED,
+                initialVisibility: VisibilityLevel.SHARED,
+                permissionUuid: ''
+            };
+        } else {
+            publicAccessFormData = {
+                visibility: VisibilityLevel.PRIVATE,
+                initialVisibility: VisibilityLevel.PRIVATE,
+                permissionUuid: ''
+            };
+        }
+
         dispatch(initialize(SHARING_PUBLIC_ACCESS_FORM_NAME, publicAccessFormData));
     };
 
@@ -209,15 +232,20 @@ const savePublicPermissionChanges = async (_: Dispatch, getState: () => RootStat
     const { user } = state.auth;
     const dialog = getDialog<SharingDialogData>(state.dialog, SHARING_DIALOG_NAME);
     if (dialog && user) {
-        const { permissionUuid, visibility } = getSharingPublicAccessFormData(state);
-        if (permissionUuid) {
-            if (visibility === VisibilityLevel.PUBLIC) {
-                await permissionService.update(permissionUuid, {
-                    name: PermissionLevel.CAN_READ
-                });
-            } else {
-                await permissionService.delete(permissionUuid);
-            }
+        const { permissionUuid, visibility, initialVisibility } = getSharingPublicAccessFormData(state);
+        // If visibility level changed, delete the previous link to public/all users.
+        // On PRIVATE this link will be deleted by saveManagementChanges
+        // so don't double delete (which would show an error dialog).
+        if (permissionUuid !== "" && visibility !== initialVisibility && visibility !== VisibilityLevel.PRIVATE) {
+            await permissionService.delete(permissionUuid);
+        }
+        if (visibility === VisibilityLevel.ALL_USERS) {
+            await permissionService.create({
+                ownerUuid: user.uuid,
+                headUuid: dialog.data.resourceUuid,
+                tailUuid: getAllUsersGroupUuid(state),
+                name: PermissionLevel.CAN_READ,
+            });
         } else if (visibility === VisibilityLevel.PUBLIC) {
             await permissionService.create({
                 ownerUuid: user.uuid,
@@ -264,7 +292,7 @@ const sendInvitations = async (_: Dispatch, getState: () => RootState, { permiss
             tailUuid: invitee.uuid,
             name: invitations.permissions
         }));
-        const changes = data.map( invitation => permissionService.create(invitation));
+        const changes = data.map(invitation => permissionService.create(invitation));
         await Promise.all(changes);
     }
 };
